mail/entreprise_gc/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.24.0
tenacity>=8.2.0
tldextract>=5.0.0
pandas>=2.0.0

# PDF parsing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Text processing
python-slugify>=8.0.0
rapidfuzz>=3.0.0

# Progress & logging
tqdm>=4.65.0

# Google Sheets integration (optional)
gspread>=5.11.0
google-auth>=2.22.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0

